// src/app/shared/models/dto/documento-tipo-processo.ts
import {DocumentoDTO} from './documento';
import {Arquivo} from "./arquivo";
import {TipoDestinatario} from "../comum/tipo-destinatario";
import {SituacaoProcesso} from "../comum/situacao-processo";
import {TipoProcesso} from "../comum/tipo-processo";
import {TipoBeneficiario} from "../comum/tipo-beneficiario";
import {EstadoCivil} from "../comum/estado-civil";
import {SimNaoEnum} from "../comum/sim-nao";
import { TipoBeneficiarioEmail } from '../comum/tipo-beneficiario-email';

export class DocumentoTipoProcesso {
    id?: number;
    idDocumento?: number;
    idDocumentoProcesso?: number;
    idGrupoDocumento?: number;
    idEstadoCivil?: number;
    idCaraterSolicitacao?: number;
    idTipoDeficiencia?: number;
    idMotivoSolicitacao?: number;
    sexo?: string;
    idadeMinima?: number;
    idadeMaxima?: number;
    valorRendaMinima?: number;
    valorRendaMaxima?: number;
    obrigatorio?: boolean;
    inativo?: string;
    dataInativacao?: Date;
    dataCadastramento?: Date;
    cpfCadastramento?: string;
    documento?: DocumentoDTO;
    arquivos?: any | Set<Arquivo> | Arquivo[];
    isToDelete?: boolean;
    idSituacaoProcesso?: number;
    idTipoProcesso?: number;
    idTipoBeneficiario?: number;
    tipoDestinatario?: TipoDestinatario;
    nomeRemetente?: string;
    emailRemetente?: string;
    nuSismail?: number;
    assunto?: string;
    texto?: string;
    matriculaCadastramento?: number;
    codigoUsuarioCadastramento?: string;
    situacaoProcesso?: SituacaoProcesso;
    tipoProcesso?: TipoProcesso;
    tipoBeneficiario?: TipoBeneficiario;
    estadoCivil?: EstadoCivil;
    tiposBeneficiariosEmail?: TipoBeneficiarioEmail[];
    tiposBeneficiario?: number[];
    tiposProcesso?: number[];
    descricaoObrigatorio?: string;
    descricaoInativo?: string;
}